Remove connect devices CTA
