+++ src/components/container/ConnectDevicesMenu/ConnectDevicesMenu.css
@@ -1,4 +1,4 @@
-.mdhui-connect-devices-menu {
+.mdhui-connect-devices-menu .mdhui-connect-devices-menu-inner {
     background: var(--mdhui-background-color-1);
     border-radius: 6px;
     margin: var(--mdhui-padding-md);
